AnimationResorucesInit Added
AnimationResorucesInit Added.
AnimTypes Moved.
Issue #755
