+++ Assets/Source/AgentAnimations/AgentAnimationTypes.cs
