+++ Assets/Source/AgentAnimations/AnimationResourcesInit.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 223dbb8ce9624d4469a4a75e63f6f793
+guid: ade159f14ed81d84d8ca86433bb5021f
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
